refactor(server-rs): decouple voice changer
